backend/requirements.txt
flask
flask-cors
mysql-connector-python>=9.0
pytest
